requirements.txt
python-telegram-bot==22.5
gspread==5.12.0
oauth2client==4.1.3
APScheduler==3.10.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

